change copy type from uint32_t to uint8_t
We can't add more field into sd_req.obj. But we can make use of 'uint32_t copies'
which effectively we only use its first 8 bytes. By squeezing copy_policy into
this room, we won't change the layout of sd_req.

For now copy_policy is just used by sheep and dog, so no need to change the QEMU
code at all.

+++ sheep/sheep_priv.h
@@ -175,7 +175,7 @@ struct vdi_iocb {
 	uint32_t snapid;
 	bool create_snapshot;
 	uint8_t copy_policy;
-	int nr_copies;
+	uint8_t nr_copies;
 	uint64_t time;
 };
